# Refactor local cache to avoid serializing entries twice