--- a/activesupport/lib/active_support/cache/null_store.rb
+++ b/activesupport/lib/active_support/cache/null_store.rb
@@ -33,10 +33,18 @@ module ActiveSupport
       end
 
       private
-        def read_entry(key, **options)
+        def read_entry(key, **s)
+          deserialize_entry(read_serialized_entry(key))
         end
 
-        def write_entry(key, entry, **options)
+        def read_serialized_entry(_key, **)
+        end
+
+        def write_entry(key, entry, **)
+          write_serialized_entry(key, serialize_entry(entry))
+        end
+
+        def write_serialized_entry(_key, _payload, **)
           true
         end