--- a/activesupport/test/cache/stores/mem_cache_store_test.rb
+++ b/activesupport/test/cache/stores/mem_cache_store_test.rb
@@ -5,26 +5,26 @@ require "active_support/cache"
 require_relative "../behaviors"
 require "dalli"
 
-# Emulates a latency on Dalli's back-end for the key latency to facilitate
-# connection pool testing.
-class SlowDalliClient < Dalli::Client
-  def get(key, options = {})
-    if /latency/.match?(key)
-      sleep 3
-      super
-    else
-      super
+class MemCacheStoreTest < ActiveSupport::TestCase
+  # Emulates a latency on Dalli's back-end for the key latency to facilitate
+  # connection pool testing.
+  class SlowDalliClient < Dalli::Client
+    def get(key, options = {})
+      if /latency/.match?(key)
+        sleep 3
+        super
+      else
+        super
+      end
     end
   end
-end
 
-class UnavailableDalliServer < Dalli::Server
-  def alive?
-    false
+  class UnavailableDalliServer < Dalli::Server
+    def alive?
+      false
+    end
   end
-end
 
-class MemCacheStoreTest < ActiveSupport::TestCase
   begin
     servers = ENV["MEMCACHE_SERVERS"] || "localhost:11211"
     ss = Dalli::Client.new(servers).stats
@@ -268,6 +268,30 @@ class MemCacheStoreTest < ActiveSupport::TestCase
     end
   end
 
+  def test_initial_object_mutation_after_fetch
+    if ActiveSupport::Cache.format_version == 6.1
+      skip "Local cache mutation can't be prevented on legacy MemCacheStore"
+    else
+      super
+    end
+  end
+
+  def test_initial_object_mutation_after_write
+    if ActiveSupport::Cache.format_version == 6.1
+      skip "Local cache mutation can't be prevented on legacy MemCacheStore"
+    else
+      super
+    end
+  end
+
+  def test_local_cache_of_read_returns_a_copy_of_the_entry
+    if ActiveSupport::Cache.format_version == 6.1
+      skip "Local cache mutation can't be prevented on legacy MemCacheStore"
+    else
+      super
+    end
+  end
+
   private
     def random_string(length)
       (0...length).map { (65 + rand(26)).chr }.join
@@ -325,6 +349,11 @@ class OptimizedMemCacheStoreTest < MemCacheStoreTest
     super
   end
 
+  def teardown
+    super
+    ActiveSupport::Cache.format_version = @previous_format
+  end
+
   def test_forward_compatibility
     previous_format = ActiveSupport::Cache.format_version
     ActiveSupport::Cache.format_version = 6.1
@@ -343,10 +372,5 @@ class OptimizedMemCacheStoreTest < MemCacheStoreTest
 
     @cache.write("foo", "bar")
     assert_equal "bar", @old_store.read("foo")
-  end
-
-  def teardown
-    super
-    ActiveSupport::Cache.format_version = @previous_format
   end
 end