@ghiculescu 2021-07-20
Hmmm, is it worth also running the benchmark against Rails 6.0 and 6.1? Since that was the initial issue in https://github.com/rails/rails/issues/42611, it would be good to see how this performs compared to those releases.