@casperisfine 2021-07-20
> Hmmm, is it worth also running the benchmark against Rails 6.0 and 6.1? Since that was the initial issue in #42611, it would be good to see how this performs compared to those releases.

Nowhere near as good. In 6.0 local cache was basically a simple hash lookup, with no deep-dup. So it's impossible to be as fast without reintroducing the leak bug.

As for 6.1, I can run the benchmark, we should be much better regardless, but it's in part because of previous PRs.