@byroot 2021-07-20
Ok, I tested it against our app, worked well except that we'll have to update https://github.com/Shopify/memcached_store to apply a similar refactoring, otherwise the local cache won't work.

I don't know if there are many popular 3rd party cache stores out there. Dalli 2.x used to ship with one, but it was removed in 3.x. Other than that I can't think of any other.